+++ agent_gantry/integrations/__init__.py
@@ -6,6 +6,10 @@ Microsoft Agent Framework, etc.
 """
 
 from agent_gantry.integrations.agent_framework_bridge import GantryToolBridge
+from agent_gantry.integrations.agent_framework_middleware import (
+    GantryApprovalMiddleware,
+    GantryObservabilityMiddleware,
+)
 from agent_gantry.integrations.framework_adapters import fetch_framework_tools
 from agent_gantry.integrations.semantic_tools import (
     SemanticToolsDecorator,
@@ -14,6 +18,8 @@ from agent_gantry.integrations.semantic_tools import (
 )
 
 __all__: list[str] = [
+    "GantryApprovalMiddleware",
+    "GantryObservabilityMiddleware",
     "GantryToolBridge",
     "SemanticToolSelector",
     "SemanticToolsDecorator",